+++ CMakeLists.txt
@@ -23,42 +23,41 @@ set(COMMON_COMPILE_OPTIONS
 	-Wno-unused-function
 	-Wno-unused-variable
 	-ffunction-sections
-	-fdata-sections 
+	-fdata-sections
 	-Wl,-gc-sections
 )
 add_definitions(
-	-DAWCHIP=AW_V5 
-	-D_GNU_SOURCE 
+	-DAWCHIP=AW_V5
+	-D_GNU_SOURCE
 	-DUSE_LOG_LIB_GLOG
 )
 
 # dependencies
 add_subdirectory(lib/lvgl)
 add_subdirectory(lib/minIni)
+add_subdirectory(lib/esp-loader)
 add_subdirectory(lib/softwinner)
 
 # application
 file(GLOB SRC_FILES_CORE   "src/core/*.c" "src/core/*.h")
 file(GLOB SRC_FILES_DRIVER "src/driver/*.c" "src/driver/*.h")
-FILE(GLOB SRC_FILES_ESP32  "src/esp32/*.c" "src/esp32/*.h")
 file(GLOB SRC_FILES_IMAGE  "src/image/*.c" "src/image/*.h")
 file(GLOB SRC_FILES_PAGE   "src/page/*.c" "src/page/*.h")
 file(GLOB SRC_FILES_BMI   "src/bmi270/*.c" "src/bmi270/*.h")
 file(GLOB SRC_FILES_WINDOW "src/window/*.c" "src/window/*.h")
 file(GLOB SRC_FILES_PLAYER "src/player/*.c" "src/player/*.h")
 
-add_executable(${PROJECT_NAME}  
-	${SRC_FILES_CORE} 
-	${SRC_FILES_DRIVER} 
-	${SRC_FILES_ESP32}
-	${SRC_FILES_IMAGE} 
-	${SRC_FILES_PAGE} 
-	${SRC_FILES_BMI} 
+add_executable(${PROJECT_NAME}
+	${SRC_FILES_CORE}
+	${SRC_FILES_DRIVER}
+	${SRC_FILES_IMAGE}
+	${SRC_FILES_PAGE}
+	${SRC_FILES_BMI}
 	${SRC_FILES_WINDOW}
 	${SRC_FILES_PLAYER}
 )
 target_compile_options(${PROJECT_NAME} PRIVATE ${COMMON_COMPILE_OPTIONS})
-target_include_directories(${PROJECT_NAME} PRIVATE 
+target_include_directories(${PROJECT_NAME} PRIVATE
 	src/
 	src/core
 	src/driver
@@ -75,11 +74,12 @@ target_link_libraries(${PROJECT_NAME} PRIVATE
 
 	lvgl
 	minIni
+	esp-loader
 	softwinner
 )
 
 # image build
-add_custom_command(TARGET ${PROJECT_NAME} POST_BUILD 
+add_custom_command(TARGET ${PROJECT_NAME} POST_BUILD
 	COMMAND size -A ${PROJECT_NAME}
 	COMMAND mkdir -p ${PROJECT_SOURCE_DIR}/out/
 	COMMAND cp ${PROJECT_NAME} ${PROJECT_SOURCE_DIR}/out/
@@ -87,4 +87,3 @@ add_custom_command(TARGET ${PROJECT_NAME} POST_BUILD
 	COMMAND cd ${PROJECT_SOURCE_DIR}/mkapp && rm -f ./ota_app/hdzgoggle_app_ota*.tar && ./mkapp_ota && ./make_fw.sh
 	COMMAND mv ${PROJECT_SOURCE_DIR}/mkapp/HDZERO_GOGGLE-*.bin ${PROJECT_SOURCE_DIR}/out/
 )
-
