+++ lib/esp-loader/src/esp_targets.c
